refactor: rename analysis workflow facade module
